+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2017-04-21 (v0.2.5)
+
+- Background shape is square, not rectangle. Doh!
+
 ## 2017-04-21 (v0.2.4)
 
 - Add support for rectangle background shape for initial avatar.